move common functions in sdlframe up to commonframe
- builds but launches to blank screen
